Safari-609.1.11.1/LayoutTests/css3/flexbox/columns-height-set-via-top-bottom.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.container {
    position: relative;
    height: 100px;
    width: 100px;
    border: 2px solid orange;
}
.flexbox {
    flex-direction: column;
    position: absolute;
    top: 0;
    right: 0;
    bottom: 0;
    left: 0;
    padding: 10px;
}
.flexbox > :nth-child(1) {
    background-color: lightblue;
}
.flexbox > :nth-child(2) {
    background-color: lightgreen;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="container">
    <div data-expected-height=100 class="flexbox column">
        <div data-expected-height=30 data-expected-width=80 style="height: 30px"></div>
        <div data-expected-height=50 data-expected-width=80 style="flex: 1;"></div>
    </div>
</div>

<div class="container">
    <div class="flexbox column wrap">
        <div data-expected-height=50 data-expected-width=40 style="height: 50px"></div>
        <div data-expected-height=80 data-expected-width=40 style="flex: 1 50px;"></div>
    </div>
</div>
</body>
</html>
